backend/requirements.txt
aiohttp==3.7.4.post0
alpha-vantage==2.3.1
aniso8601==9.0.1
appdirs==1.4.3
astroid==2.5.3
async-timeout==3.0.1
attrs==20.3.0
bcrypt==3.2.0
better-profanity==0.7.0
black==20.8b1
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.6
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
Flask-Cors==3.0.10
flask-restx==0.3.0
html5lib==1.0.1
idna==2.8
iexfinance==0.5.0
ipaddr==2.2.0
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
lazy-object-proxy==1.6.0
lockfile==0.12.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
multidict==5.1.0
mypy-extensions==0.4.3
numpy==1.20.1
packaging==20.3
pandas==1.2.3
pathspec==0.8.1
pep517==0.8.2
progress==1.5
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pylint==2.7.4
pylint-venv==2.1.1
pyparsing==2.4.6
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytoml==0.1.21
pytz==2021.1
regex==2020.11.13
requests==2.22.0
retrying==1.3.3
simplejson==3.17.2
six==1.14.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
